Final version for the presentation
This version still have some bugs during we rotating our skybox. There
will be certain faces will disappear during rotation. It was because the
vertex coordinate is abnormal to cause this problem.

More, this version can only change the asc and skybox texture by hand to
modify the Gz.h and Application5.c. During changing the model or the
mode, you have some configuration need to be changed too, such as model
scale and rotation. And when you changed the mode, you will need to
modified the text_fun to reverse the texture of skybox.